Made code to be c89 compliant

+++ bench.c
@@ -34,7 +34,6 @@
         } \
  \
         ts(start)
-
 #define BENCH_END() \
     ts(finish); \
     ts_add_diff(taken, start, finish); \
@@ -50,14 +49,14 @@
  */
 
 static void bench_report(
-    void*   taken,  /* time taken copying data */
-    size_t  copied) /* number of bytes copied */
+    void*   taken,  /* time taken on data copying */
+    float   copied) /* number of bytes copied */
 {
     /*~~~~~~~~~~~~~~~~~~~~~~~~*/
     struct jedec    jd_bps;     /* bytes per second in jedec format */
     struct jedec    jd_copied;  /* number of bytes copied in jedec format */
     unsigned long   ns;         /* time taken copying data in nanoseconds */
-    double          bps;        /* bytes per second rate */
+    float           bps;        /* bytes per second rate */
     /*~~~~~~~~~~~~~~~~~~~~~~~~*/
 
     if ((ns = ts2ns(taken)) == 0)
@@ -65,14 +64,18 @@ static void bench_report(
         ns = 1;
     }
 
-    bps = (double)copied / ns;
+    bps = copied / ns;
     bps *= 1000000;
 
     bytes2jedec(bps, &jd_bps);
     bytes2jedec(copied, &jd_copied);
 
-    printf("copied %5zu %cB, in %5lu ns, rate %5zu %cB/s\n", jd_copied.val,
-           jd_copied.pre, ns, jd_bps.val, jd_bps.pre);
+    printf("copied %5lu %cB, in %5lu ns, rate %5lu %cB/s\n",
+           jd_copied.val,
+           jd_copied.pre,
+           ns,
+           jd_bps.val,
+           jd_bps.pre);
 }
 
 /*
@@ -105,7 +108,7 @@ static void bench_fill_random(
  -------------------------------------------------------------------------------
  */
 
-int bench(void* dst, /* destination pointer */ void* src)   /* source pointer */
+int bench(void* dst, /* */ void* src)   /* */
 {
     /*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
     struct jedec    jd_block_size;      /* block size in jedec format */
@@ -113,7 +116,7 @@ int bench(void* dst, /* destination pointer */ void* src)   /* source pointer */
     void*           start;              /* timer indicating benchmark start */
     void*           finish;             /* timer indicating benchmark finish */
     void*           taken;              /* timer for time taken on benchmark */
-    size_t          bytes_copied;       /* bytes copied in * iteration */
+    float           bytes_copied;       /* bytes copied in * iteration */
     size_t          loops;              /* loops needed to copy requested bytes */
     size_t          i;                  /* iterator for loop */
     size_t          j;                  /* iterator for loop */
@@ -129,17 +132,20 @@ int bench(void* dst, /* destination pointer */ void* src)   /* source pointer */
     bytes2jedec(opts.block_size, &jd_block_size);
     bytes2jedec(opts.report_intvl, &jd_report_intvl);
 
-    printf("block size: %zu %cB, report every %zu %cB, iterations %zu\n",
-           jd_block_size.val, jd_block_size.pre, jd_report_intvl.val,
-           jd_report_intvl.pre, opts.num_intvl);
+    printf("block size: %lu %cB, report every %lu %cB, iterations %lu\n",
+           jd_block_size.val,
+           jd_block_size.pre,
+           jd_report_intvl.val,
+           jd_report_intvl.pre,
+           opts.num_intvl);
 
     /*
      * for systems that uses optimistic memory allocation (like linux) dst
-     * and src are not really allocated just yet. dst and src will be
+     * and src may not really allocated just yet. dst and src will be
      * allocated when we first access them. This causes first memory copy
      * iteration to take much longer time causing program to show wrong
      * transfer rate. To prevent this behaviour we do a simple memory copy
-     * here, so dst and src are allocated.
+     * here, so dst an
      */
 
     memcpy(dst, src, opts.block_size);
@@ -151,14 +157,12 @@ int bench(void* dst, /* destination pointer */ void* src)   /* source pointer */
         switch (opts.method)
         {
         case METHOD_MEMCPY:
-
             BENCH_START();
             memcpy(dst, src, opts.block_size);
             BENCH_END();
             break;
 
         case METHOD_BBB:
-
             BENCH_START();
 
             for (k = 0; k != opts.block_size; ++k)
@@ -181,7 +185,7 @@ int bench(void* dst, /* destination pointer */ void* src)   /* source pointer */
             assert(0 && "test method not supported, should not get here");
         }
 
-        bytes_copied = j * opts.block_size;
+        bytes_copied = (float)j * opts.block_size;
         bench_report(taken, bytes_copied);
     }
